♻️ refactor: extract phase labels from ModelDownloadHostView
Split `currentPhaseLabel` + `phaseDisplayName` into a sibling
extension file (`+PhaseLabels.swift`), mirroring the existing
`+Routing.swift` / `+StateFallbacks.swift` split. After main's
#262 added the `.eventInject` switch case (1 line), the host
file hit the `file_length` 400 limit; the extraction drops it
to 374. Visibility raised from `private` → internal (default)
since extension methods can't share `private` across files —
both methods stay only meaningfully callable from the host
view (no new public surface).
